+++ src/main/java/gift/dto/kakao/KakaoErrorResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.kakao;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
