+++ src/main/java/gift/dto/kakao/KakaoMemberLoginResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.kakao;
 
 public record KakaoMemberLoginResponseDto(String token) {
 }
