+++ src/main/java/gift/dto/kakao/KakaoTokenResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.kakao;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
