+++ src/main/java/gift/dto/member/MemberLoginRequestDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.member;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
