+++ src/main/java/gift/dto/member/MemberLoginResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.member;
 
 public record  MemberLoginResponseDto (String token) {
 }
