+++ src/main/java/gift/dto/member/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.member;
 
 import jakarta.validation.constraints.NotBlank;
