+++ src/main/java/gift/dto/member/MemberResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.member;
 
 public record MemberResponseDto(Long id, Long kakaoId, String nickname) {
 }
