+++ src/main/java/gift/dto/option/OptionRequestDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.option;
 
 import jakarta.validation.constraints.*;
