+++ src/main/java/gift/dto/product/ProductResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.product;
 
 public record ProductResponseDto(
         Long id, String name, Integer price, String imageUrl) {
