+++ src/main/java/gift/dto/wish/WishCreateResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.wish;
 
 public record WishCreateResponseDto(Long id, Long memberId, Long productId) {
 }
