+++ src/main/java/gift/dto/wish/WishRequestDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.wish;
 
 public record WishRequestDto (Long productId) {
 }
