+++ src/main/java/gift/dto/wish/WishResponseDto.java
@@ -1,4 +1,6 @@
-package gift.dto;
+package gift.dto.wish;
+
+import gift.dto.product.ProductResponseDto;
 
 public record WishResponseDto(
         Long id, ProductResponseDto product) {
