+++ src/main/java/gift/service/WishService.java
@@ -1,8 +1,8 @@
 package gift.service;
 
-import gift.dto.ProductResponseDto;
-import gift.dto.WishCreateResponseDto;
-import gift.dto.WishResponseDto;
+import gift.dto.product.ProductResponseDto;
+import gift.dto.wish.WishCreateResponseDto;
+import gift.dto.wish.WishResponseDto;
 import gift.entity.Member;
 import gift.entity.Product;
 import gift.entity.Wish;
