+++ src/main/resources/application.properties
@@ -12,3 +12,11 @@ spring.jpa.show-sql=true
 spring.jpa.open-in-view=false
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.defer-datasource-initialization=true
+
+kakao.client-id=af6e80383ad6bbfe5fe26ed030845191
+kakao.redirect-uri=http://localhost:8080/oauth/kakao/callback
+kakao.token-url=https://kauth.kakao.com/oauth/token
+kakao.send-message-url=https://kapi.kakao.com/v2/api/talk/memo/default/send
+kakao.user-info-url=https://kapi.kakao.com/v2/user/me
+kakao.redirect-message-web-url=http://localhost:8080/orders
+kakao.redirect-message-mobile-url=http://localhost:8080/orders
